More ignored environment variables; AB_NO_SANDBOX can be used to turn the sandbox off; shart command lines are inlined.

+++ build/ab.py
@@ -55,6 +55,7 @@ class Environment(types.SimpleNamespace):
 
 
 G = Environment()
+G.setdefault("AB_NO_SANDBOX", "")
 
 
 class PathFinderImpl(PathFinder):
@@ -558,19 +559,23 @@ def emit_rule(self, ins, outs, cmds=[], label=None):
         os.makedirs(self.dir, exist_ok=True)
         rule = []
 
-        sandbox = join(self.dir, "sandbox")
-        emit(f"rm -rf {sandbox}", into=rule)
-        emit(
-            f"{G.PYTHON} build/_sandbox.py --link -s", sandbox, *fins, into=rule
-        )
-        for c in cmds:
-            emit(f"(cd {sandbox} &&", c, ")", into=rule)
-        emit(
-            f"{G.PYTHON} build/_sandbox.py --export -s",
-            sandbox,
-            *fouts,
-            into=rule,
-        )
+        if G.AB_NO_SANDBOX:
+            sandbox = join(self.dir, "sandbox")
+            emit(f"rm -rf {sandbox}", into=rule)
+            emit(
+                f"{G.PYTHON} build/_sandbox.py --link -s", sandbox, *fins, into=rule
+            )
+            for c in cmds:
+                emit(f"(cd {sandbox} &&", c, ")", into=rule)
+            emit(
+                f"{G.PYTHON} build/_sandbox.py --export -s",
+                sandbox,
+                *fouts,
+                into=rule,
+            )
+        else:
+            for c in cmds:
+                emit("(", c, ")", into=rule)
 
         ruletext = "".join(rule)
         if len(ruletext) > 7000:
@@ -581,13 +586,13 @@ def emit_rule(self, ins, outs, cmds=[], label=None):
                 fp.write("set -e\n")
                 fp.write(ruletext)
 
-            emit("build", *fouts, ":rule", *fins, rulef)
+            emit("build", *fouts, ":rule", *fins)
             emit(" command=sh", rulef)
         else:
             emit("build", *fouts, ":rule", *fins)
             emit(
                 " command=",
-                "&&".join([s.strip() for s in rule]).replace("$", "$$"),
+                " && ".join([s.strip() for s in rule]).replace("$", "$$"),
             )
         if label:
             emit(" description=", label)
